Exported and renamed a few packages in the Netatmo binding. See openhab/static-code-analysis#100
