+++ addons/binding/org.openhab.binding.netatmo/src/main/java/org/openhab/binding/netatmo/handler/homecoach/NAHealthyHomeCoachHandler.java
@@ -13,11 +13,11 @@ import static org.openhab.binding.netatmo.NetatmoBindingConstants.*;
 import org.eclipse.smarthome.core.library.types.DecimalType;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.types.State;
-import org.openhab.binding.netatmo.config.NetatmoDeviceConfiguration;
 import org.openhab.binding.netatmo.handler.NetatmoDeviceHandler;
 import org.openhab.binding.netatmo.internal.ChannelTypeUtils;
 import org.openhab.binding.netatmo.internal.NADeviceAdapter;
 import org.openhab.binding.netatmo.internal.NAHealthyHomeCoachAdapter;
+import org.openhab.binding.netatmo.internal.config.NetatmoDeviceConfiguration;
 
 import io.swagger.client.model.NADashboardData;
 import io.swagger.client.model.NAHealthyHomeCoachDataBody;
