+++ addons/binding/org.openhab.binding.netatmo/src/main/java/org/openhab/binding/netatmo/internal/config/NetatmoModuleConfiguration.java
@@ -6,21 +6,19 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.netatmo.config;
+package org.openhab.binding.netatmo.internal.config;
 
 /**
- * The {@link NetatmoThingConfiguration} is the base class for configuration
- * information held by devices and modules
+ * The {@link NetatmoModuleConfiguration} is responsible for holding configuration
+ * informations needed to access a Netatmo Module (that depends upon a device)
  *
  * @author Gaël L'hopital - Initial contribution
  */
-public class NetatmoThingConfiguration {
+public class NetatmoModuleConfiguration extends NetatmoThingConfiguration {
+    private String parentId;
 
-    private String equipmentId;
-
-    public String getEquipmentId() {
+    public String getParentId() {
         // Bug #3891 : Netatmo API only works with lower case device/module ids
-        return equipmentId.toLowerCase();
+        return parentId.toLowerCase();
     }
-
 }
